+++ src/main/java/com/gxdxx/instagram/service/user/UserProfileQueryService.java
@@ -1,0 +1,33 @@
+package com.gxdxx.instagram.service.user;
+
+import com.gxdxx.instagram.dto.response.UserProfileResponse;
+import com.gxdxx.instagram.entity.User;
+import com.gxdxx.instagram.exception.UserNotFoundException;
+import com.gxdxx.instagram.repository.FollowRepository;
+import com.gxdxx.instagram.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Transactional
+@RequiredArgsConstructor
+@Service
+public class UserProfileQueryService {
+
+    private final UserRepository userRepository;
+    private final FollowRepository followRepository;
+
+    @Transactional(readOnly = true)
+    public UserProfileResponse findUserProfile(String nickname) {
+        User user = getUserFromNickname(nickname);
+        Long followerCount = followRepository.countByFollowing(user);
+        Long followingCount = followRepository.countByFollower(user);
+        return UserProfileResponse.of(user.getNickname(), user.getProfileImageUrl(), followerCount, followingCount);
+    }
+
+    private User getUserFromNickname(String nickname) {
+        return userRepository.findByNickname(nickname)
+                .orElseThrow(UserNotFoundException::new);
+    }
+
+}
